Fix bug where a JVM restart with sapling parcels
Drop a copy of the parcel manager managed itemstack
rather than the actual itemstack itself.  When the
actual itemstack is dropped it gets emptied when the
player picks up the item, causing a JVM restart to
be necessary for a player to get a second sapling
of the same type from the ancient parcels.